+++ bootmgr-rs-ratatui/src/app.rs
@@ -168,12 +168,6 @@ impl App {
         Ok(Some(handle))
     }
 
-    /// Drops the [`Terminal`], and consumes `self`.
-    pub fn close(self, terminal: Terminal<UefiBackend>) {
-        let _ = self;
-        drop(terminal);
-    }
-
     /// Initializes the state of the terminal and events.
     ///
     /// # Errors